# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FoggyBrain HTTP API",
    "version": "1.0.0",
    "description": "Authoritative HTTP shapes, routes, and transport rules, generated from src/shared.ts and scripts/openapi.ts. See CONTRACT.md for domain semantics. Trusted local tool, not an authenticated multi-user service: bind to loopback and do not expose publicly. Host/Origin and cross-site request checks are enforced, but are not user authentication. Credentials stay on the server: dedicated sync uses only FOGGY_SYNC_TOKEN; explicit github mode reuses GH_TOKEN, GITHUB_TOKEN, or gh auth token. Cloud sync requires Contents read/write on the configured private repository; PR polling credentials should be read-only. JSON requests are limited to 256kb. POST/PATCH/PUT require application/json except an absent GitHub sync body. All errors are JSON; malformed JSON/URLs return 400, rejected local request headers 403, oversized bodies 413, unsupported media type/encoding 415, and unexpected failures 500."
  },
  "servers": [{ "url": "http://127.0.0.1:4173" }],
...
  "paths": {
...
    "/api/workspaces": {
      "get": {
        "operationId": "listWorkspaces",
        "summary": "List local workspace registrations",
        "description": "Unscoped management endpoint. See CONTRACT.md for state-dependent validation and semantics.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "Cache-Control": {
                "description": "API responses are not cached, including credential-backed discovery.",
                "schema": { "type": "string", "const": "no-store" }
              }
            },
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/WorkspaceList" } }
            }
          },
          "400": {
            "description": "Malformed JSON/URL, unsupported fields, invalid input, or missing confirmation.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "403": {
            "description": "Host, Origin, or cross-site request rejected by the local-tool boundary.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "413": {
            "description": "JSON body exceeds the 256kb limit.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "415": {
            "description": "Unsupported Content-Type, JSON charset, or content encoding.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "500": {
            "description": "Unexpected internal server error; details are not exposed.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        },
        "x-query-policy": "undeclared-ignored"
      },
      "post": {
        "operationId": "createWorkspace",
        "summary": "Create a local or cloud workspace",
        "description": "Saves configuration only; does not fetch or publish remote state. The first workspace becomes the default. Unscoped management endpoint. See CONTRACT.md for state-dependent validation and semantics.",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created.",
            "headers": {
              "Cache-Control": {
                "description": "API responses are not cached, including credential-backed discovery.",
                "schema": { "type": "string", "const": "no-store" }
              }
            },
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Workspace" } }
            }
          },
          "400": {
            "description": "Malformed JSON/URL, unsupported fields, invalid input, or missing confirmation.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "403": {
            "description": "Host, Origin, or cross-site request rejected by the local-tool boundary.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "409": {
            "description": "Graph conflict, workspace limit, active sync, or stale preview; review state before retrying.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "413": {
            "description": "JSON body exceeds the 256kb limit.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "415": {
            "description": "Unsupported Content-Type, JSON charset, or content encoding.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "500": {
            "description": "Unexpected internal server error; details are not exposed.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "503": {
            "description": "Selected credentials/configuration unavailable or workspace service stopped.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        },
        "x-query-policy": "undeclared-ignored",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateWorkspaceInput" }
            }
          }
        }
      }
    },
    "/api/workspaces/repositories": {
      "get": {
        "operationId": "listWorkspaceRepositories",
        "summary": "Discover private user-owned repositories",
        "description": "Read-only server-authenticated discovery. Omitted credential defaults to dedicated. No automatic credential fallback; no organization or collaborator repositories. Rejects duplicate or unsupported query fields and incomplete upstream results. Unscoped management endpoint. See CONTRACT.md for state-dependent validation and semantics.",
        "parameters": [
          {
            "name": "credential",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["dedicated", "github"],
              "description": "Dedicated uses only server FOGGY_SYNC_TOKEN (default); github explicitly reuses server GH_TOKEN, GITHUB_TOKEN, or gh auth token. Never accepts browser credentials."
            },
            "description": "Dedicated uses only server FOGGY_SYNC_TOKEN (default); github explicitly reuses server GH_TOKEN, GITHUB_TOKEN, or gh auth token. Never accepts browser credentials."
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "Cache-Control": {
                "description": "API responses are not cached, including credential-backed discovery.",
                "schema": { "type": "string", "const": "no-store" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WorkspaceRepositories" }
              }
            }
          },
          "400": {
            "description": "Malformed JSON/URL, unsupported fields, invalid input, or missing confirmation.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "403": {
            "description": "Host, Origin, or cross-site request rejected by the local-tool boundary.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "413": {
            "description": "JSON body exceeds the 256kb limit.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "415": {
            "description": "Unsupported Content-Type, JSON charset, or content encoding.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "500": {
            "description": "Unexpected internal server error; details are not exposed.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "502": {
            "description": "Sanitized upstream GitHub failure.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "503": {
            "description": "Selected credentials/configuration unavailable or workspace service stopped.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        },
        "x-query-policy": "declared-only"
      }
    },
...
    "/api/workspaces/{id}/removal-preview": {
      "get": {
        "operationId": "previewWorkspaceRemoval",
        "summary": "Preview local workspace removal",
        "description": "Local-only impact and revision fingerprint; no remote request. Review before confirming removal. Unscoped management endpoint. See CONTRACT.md for state-dependent validation and semantics.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Server-returned resource ID; relationship removal uses the relationship ID, not a task ID."
          }
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "headers": {
              "Cache-Control": {
                "description": "API responses are not cached, including credential-backed discovery.",
                "schema": { "type": "string", "const": "no-store" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WorkspaceRemovalPreview" }
              }
            }
          },
          "400": {
            "description": "Malformed JSON/URL, unsupported fields, invalid input, or missing confirmation.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "403": {
            "description": "Host, Origin, or cross-site request rejected by the local-tool boundary.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "404": {
            "description": "Route, workspace, task, or relationship not found; no default workspace exists.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "409": {
            "description": "Graph conflict, workspace limit, active sync, or stale preview; review state before retrying.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "413": {
            "description": "JSON body exceeds the 256kb limit.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "415": {
            "description": "Unsupported Content-Type, JSON charset, or content encoding.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "500": {
            "description": "Unexpected internal server error; details are not exposed.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          },
          "503": {
            "description": "Selected credentials/configuration unavailable or workspace service stopped.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
            }
          }
        },
        "x-query-policy": "declared-only"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SyncTarget": {
        "type": "object",
        "properties": {
          "repo": { "$ref": "#/components/schemas/SyncRepository" },
          "branch": { "$ref": "#/components/schemas/SyncRef" },
          "path": { "$ref": "#/components/schemas/SyncRef" }
        },
        "required": ["repo", "branch", "path"],
        "additionalProperties": false
      },
      "Workspace": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "type": { "type": "string", "enum": ["local", "cloud"] },
          "target": {
            "anyOf": [{ "$ref": "#/components/schemas/SyncTarget" }, { "type": "null" }]
          },
          "credential": { "type": ["string", "null"], "enum": ["dedicated", "github", null] }
        },
        "required": ["id", "name", "type", "target", "credential"],
        "additionalProperties": false,
        "allOf": [
          {
            "if": { "properties": { "type": { "const": "local" } }, "required": ["type"] },
            "then": {
              "properties": { "target": { "type": "null" }, "credential": { "type": "null" } }
            },
            "else": {
              "properties": {
                "target": { "not": { "type": "null" } },
                "credential": { "not": { "type": "null" } }
              }
            }
          }
        ]
      },
      "WorkspaceList": {
        "type": "object",
        "properties": {
          "workspaces": { "type": "array", "items": { "$ref": "#/components/schemas/Workspace" } },
          "defaultWorkspaceId": { "type": ["string", "null"] },
          "limit": { "type": "number" }
        },
        "required": ["workspaces", "defaultWorkspaceId", "limit"],
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"],
        "additionalProperties": false
      },
...
    }
  }
}